Chore: Move Dummy Properties to TestFile
