+++ feature/explore/src/test/kotlin/kr/co/explore/ExploreViewModelTest.kt
@@ -16,9 +16,11 @@ import kr.co.data.repository.RecentRepository
 import kr.co.model.ExploreSideEffect
 import kr.co.model.ExploreUiIntent
 import kr.co.model.FileInfo
+import kr.co.model.FileInfo.Type.PDF
 import kr.co.util.FileManagerImpl
 import org.junit.Before
 import org.junit.Test
+import java.time.LocalDateTime
 import kotlin.test.assertEquals
 
 class ExploreViewModelTest {
@@ -28,6 +30,30 @@ class ExploreViewModelTest {
     private lateinit var fileManagerImpl: FileManagerImpl
     private val testDispatcher = StandardTestDispatcher()
     private val testScope = TestScope(testDispatcher)
+
+    companion object {
+        val PDF_DUMMY = FileInfo(
+            name = "DUMMY.pdf",
+            path = "",
+            type = PDF,
+            isDirectory = false,
+            isHidden = false,
+            size = 0,
+            createdAt = LocalDateTime.now(),
+            lastModified = LocalDateTime.now()
+        )
+
+        val FOLDER_DUMMY = FileInfo(
+            name = "DUMMY",
+            path = "",
+            type = PDF,
+            isDirectory = true,
+            isHidden = false,
+            size = 0,
+            createdAt = LocalDateTime.now(),
+            lastModified = LocalDateTime.now()
+        )
+    }
 
     @Before
     fun setup() {
@@ -41,11 +67,11 @@ class ExploreViewModelTest {
     fun `Init intent 경로를 통한 파일 초기화`() = testScope.runTest {
         val path = "/path"
         val folders = listOf(
-            FileInfo.FOLDER_DUMMY
+            FOLDER_DUMMY
         )
         val files = listOf(
-            FileInfo.PDF_DUMMY,
-            FileInfo.PDF_DUMMY
+            PDF_DUMMY,
+            PDF_DUMMY
         )
 
         coEvery { fileManagerImpl.readPDFOrDirectory(path) } returns files + folders
@@ -65,7 +91,7 @@ class ExploreViewModelTest {
 
     @Test
     fun `ClickFile Repository에 insert 후 PDF 화면으로 이동`() = testScope.runTest {
-        val file = FileInfo.PDF_DUMMY
+        val file = PDF_DUMMY
 
         coEvery { recentRepository.insert(file) } just Runs
 
@@ -83,7 +109,7 @@ class ExploreViewModelTest {
 
     @Test
     fun `ClickFolder Folder 화면으로 이동`() = testScope.runTest {
-        val folder = FileInfo.FOLDER_DUMMY
+        val folder = FOLDER_DUMMY
 
         viewModel.handleIntent(ExploreUiIntent.ClickFolder(folder))
